+++ MLibs/version.properties
@@ -1,4 +1,4 @@
-#Thu Sep 07 18:06:38 CST 2023
-buildTime=0907180638
-versionName=1.0.643
-versionCode=643
+#Thu Sep 07 21:02:33 CST 2023
+versionName=1.0.650
+buildTime=0907210233
+versionCode=650
