+++ app/src/main/java/cn/mvp/MainActivity.java
@@ -17,14 +17,12 @@ import com.king.zxing.Intents;
 import com.king.zxing.util.CodeUtils;
 
 import java.io.IOException;
-import java.net.DatagramSocket;
 
 import cn.mvp.acty.BaseActivity;
 import cn.mvp.acty.ElectricQuantityActivity;
 import cn.mvp.acty.zfb.ZfbActivity;
 import cn.mvp.chat.ChatActivity;
 import cn.mvp.chat1.Chat1Activity;
-import cn.mvp.chat3.UdpUtils;
 import cn.mvp.global.Constant;
 import cn.mvp.mlibs.utils.ClipboardUtils;
 import cn.mvp.mlibs.utils.DeviceUtils;
@@ -37,9 +35,6 @@ public class MainActivity extends BaseActivity implements View.OnClickListener {
 
     private boolean isRefuse;
     private TextView mTv;
-    private DatagramSocket mClientSocket;
-    private byte[] mReceiveData;
-
 
     private void handleSendText(Intent intent) {
         String sharedText = intent.getStringExtra(Intent.EXTRA_TEXT);
@@ -129,7 +124,7 @@ public class MainActivity extends BaseActivity implements View.OnClickListener {
             mTv.setText(DeviceUtils.getDeviceInfo());//获取手机基本信息
         });
         findViewById(R.id.main_btn_base_tv_test).setOnClickListener(v -> {
-            TestActivity.open(MainActivity.this);//测试按钮
+//            TestActivity.open(MainActivity.this);//测试按钮
         });
 
         findViewById(R.id.main_btn_test_page).setOnClickListener(v -> {
